backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ezdxf==1.3.3
neo4j==5.20.0
pymupdf==1.24.3
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1